webapp/requirements.txt
bokeh==3.3.0
cassandra-driver==3.28.0
cqlsh==6.1.2
Django==4.2.6
django-cassandra-engine==1.8.0
django-filter==23.5
djangorestframework==3.14.0
hvplot==0.9.0
pandas==2.1.1
psycopg2-binary==2.9.9
yfinance==0.2.31
